modifie mappers est controller

+++ src/main/java/com/ouadia/rovista1/repositories/CategorieRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface CategorieRepository extends JpaRepository<Categorie,Integer> {
+public interface CategorieRepository extends JpaRepository<Categorie,Long> {
     Categorie findByNom(String nom);
 
 }
